requirements.txt
# Core dependencies for washdb-bot
sqlalchemy>=2.0.0
psycopg[binary]>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
tldextract>=5.0.0

# Scheduling
APScheduler>=3.10.0

# Headless browser for anti-blocking
playwright>=1.40.0
